MAster branch updated to include all the necessary lib, dll, and header files.
The structure for this is the standard single include directory and the build dependant lib directories (x86/x64).

We currently don't have CMake set up to include and link our files so you will have to do this manually. If you are unsure of this process please let me know and I will guide you through it. Sam covered it a bit as a general proces in his last lecture, but for the time being we will need only a few of the files so we don't need to link them all.