I think threads are working so that collision detection is working for 2 entities
The idea is that the CycleThroughEntities() method now takes in a string that corresponds with the name/id of a player entity. It then checks whether or not the related entity is colliding with any of the other entities in the scene.

It's kinda hard to tell, and it messes up the console output, but I think it's working so that we can now push back a thread that calls the CycleThroughEntities() method, one passing the "ob1" string (temporarily, should change to player1 and player2), and the other passing "ob2."

So, the first thread checks the corresponding "ob1" model against all other entities, and the other thread simultaneously checks to see if "ob2" is colliding with anything.