Copied Jeremy's updated CMakeLists.txt and added music lib
Also stopped it from copying the GLM .cpp into source files cause that gives LNK errors. Also set it up to copy the contents of the sounds folder, and created a folder of DLLs to match up with Jeremy's work with CMake. Finally, I made it work in release mode.