Right I'm done with sound
Obviously it needs to know if it's in a menu or if it's in a game and we need real audio files and the shooting sound will need hooked up to shots, and maybe we can control the volume through the controllers too but idk about that. Atm it's working with M for mute, + for normal volume and - for low. Not much difference between normal and low so might change it to medium or high but then it kinda gets too loud so I dunno